requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.85.0
anyio==4.12.1
async-timeout==4.0.3
attrs==25.4.0
banks==2.4.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.4.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
eval_type_backport==0.2.2
exceptiongroup==1.3.1
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.73.0
greenlet==3.2.5
griffe==1.14.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.1
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
langchain==0.3.28
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.16
llama-index-core==0.12.52.post1
llama-index-embeddings-openai==0.5.2
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.3
llama-index-llms-openai==0.6.26
llama-index-readers-file==0.5.6
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-chroma==0.4.1
llama-index-workflows==1.3.0
llama-parse==0.6.54
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.2
numpy==2.0.2
oauthlib==3.3.1
onnxruntime==1.19.2
openai==2.29.0
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==21.0.0
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.5
pypdf==6.9.1
PyPika==0.51.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
streamlit==1.50.0
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tinytag==2.2.1
tokenizers==0.22.2
toml==0.10.2
tomli==2.4.0
tornado==6.5.5
tqdm==4.67.3
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
